+++ rand/randutil.cpp
@@ -3,7 +3,7 @@
 
 uint rand(uint from, uint to) {
 	if (from < to) {
-		return rand() % ((to - from) + 1) + from;
+		return (uint)rand() % ((to - from) + 1) + from;
 	}
 	return 0;
 }
